+++ index.js
@@ -1,3 +1,0 @@
-const Enveloop = require('./lib/enveloop')
-
-module.exports = Enveloop