+++ src/lib/enveloop.ts
@@ -1,8 +1,41 @@
-const fetch = require('node-fetch')
-const { version } = require('../package')
+import fetch from 'node-fetch'
+import { version } from '../../package.json'
+
+type EnveloopOptions = {
+  apiKey: string
+  apiHost?: string
+  ssl?: boolean
+}
+
+type EnveloopMessageOptions = {
+  template?: string
+  html?: string
+  to: string
+  from: string
+  subject: string
+  templateVariables?: object
+}
+
+type EnveloopTemplateOptions = {
+  template: string
+}
+
+type EnveloopMessageResponse = {
+  to: string
+  from: string
+  body: string
+}
+
+type EnveloopTemplateResponse = {
+  templateVariables: [string]
+  body: string
+}
 
 class Enveloop {
-  constructor ({ apiKey, apiHost = 'api.enveloop.com', ssl = true }) {
+  private endpoint: string
+  private apiKey: string
+
+  constructor ({ apiKey, apiHost = 'api.enveloop.com', ssl = true }: EnveloopOptions) {
     if (!apiKey) {
       throw new Error('apiKey value must be defined!')
     }
@@ -11,8 +44,8 @@ class Enveloop {
     this.apiKey = apiKey
   }
 
-  async sendMessage ({ template, to, from, subject, templateVariables = {} }) {
-    const data = { to, from, subject, template, templateVariables }
+  async sendMessage ({ template, html, to, from, subject, templateVariables = {} }: EnveloopMessageOptions): Promise<{ status: number, message: EnveloopMessageResponse }> {
+    const data = { template, html, to, from, subject, templateVariables }
 
     const response = await fetch(`${this.endpoint}/messages`, {
       method: "POST",
@@ -32,7 +65,7 @@ class Enveloop {
     return { status: response.status, message }
   }
   
-  async templateInfo ({ template }) {
+  async templateInfo ({ template }: EnveloopTemplateOptions): Promise<{ status: number, template: EnveloopTemplateResponse }> {
     const response = await fetch(`${this.endpoint}/templates/${template}`, {
       method: "GET",
       headers: {
@@ -46,8 +79,10 @@ class Enveloop {
       throw new Error(templateInfo.error)
     }
 
+    console.log({templateInfo})
+
     return { status: response.status, template: templateInfo }
   }
 }
 
-module.exports = Enveloop
+export { Enveloop }